feat(player): Position saving

+++ server/player/index.ts
@@ -14,12 +14,26 @@ onNet("test-env:playerLoaded", async () => {
               identifier: player.identifier,
               phoneNumber: player.phone_number
           })
+          emitNet("test-env:spawnPlayer", src, player.position)
       }
     })
 })
 
 on("playerDropped", async () => {
     const src = global.source
-    Logger.info(`Player disconnected source: ${src}`)
-    PlayerService.removePlayer(src)
+    const Player = PlayerService.getPlayer(src)
+    if (!Player) {
+        PlayerService.removePlayer(src)
+        return
+    }
+
+    Logger.info(`Player disconnected: ${src}: ${Player.name}`)
+
+    const [x, y, z] = GetEntityCoords(GetPlayerPed(src.toString()))
+    Logger.debug(`Player[${src}] coords: [${x}, ${y}. ${z}]`)
+
+    Player.position = {x, y, z}
+    PlayerService.updatePlayer(Player).then(() => {
+        PlayerService.removePlayer(src)
+    })
 })
